Refractor files of the QrCameraScanner and ItemInformation (#111)
