+++ app/src/androidTest/java/com/github/brugapp/brug/ItemInformationActivityTest.kt
@@ -9,6 +9,7 @@ import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.brugapp.brug.model.Item
+import com.github.brugapp.brug.ui.ItemInformationActivity
 import org.junit.Test
 import org.junit.runner.RunWith
