+++ app/src/androidTest/java/com/github/brugapp/brug/ItemsMenuActivityTest.kt
@@ -17,7 +17,9 @@ import org.hamcrest.Matchers.allOf
 import androidx.test.uiautomator.UiScrollable
 import androidx.test.uiautomator.UiSelector
 import com.github.brugapp.brug.ui.ChatMenuActivity
+import com.github.brugapp.brug.ui.ItemInformationActivity
 import com.github.brugapp.brug.ui.ItemsMenuActivity
+import com.github.brugapp.brug.ui.QrCodeScannerActivity
 import com.github.brugapp.brug.view_model.ListViewHolder
 import org.hamcrest.core.IsEqual
 import org.junit.After
