+++ app/src/androidTest/java/com/github/brugapp/brug/QrCodeScannerActivityTest.kt
@@ -10,6 +10,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withHint
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.brugapp.brug.ui.QrCodeScannerActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
