+++ app/src/androidTest/java/com/github/brugapp/brug/SettingsTest.kt
@@ -1,8 +1,5 @@
 package com.github.brugapp.brug
 
-import android.content.Intent
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.espresso.Espresso
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions
