+++ app/src/main/java/com/github/brugapp/brug/MainActivity.kt
@@ -6,11 +6,7 @@ import android.view.Menu
 import android.view.MenuItem
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
-import com.github.brugapp.brug.ui.ChatActivity
-import com.github.brugapp.brug.ui.ItemsMenuActivity
-import com.github.brugapp.brug.ui.RegisterUserActivity
-import com.github.brugapp.brug.ui.SettingsActivity
-import com.github.brugapp.brug.ui.SignInActivity
+import com.github.brugapp.brug.ui.*
 
 class MainActivity : AppCompatActivity() {
