+++ app/src/main/res/layout/activity_item_information.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ItemInformationActivity">
+    tools:context=".ui.ItemInformationActivity">
 
     <RelativeLayout
         android:id="@+id/rellay1"
