+++ app/src/main/res/layout/activity_qr_code_scanner.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".QrCodeScannerActivity">
+    tools:context=".ui.QrCodeScannerActivity">
 
     <EditText
         android:id="@+id/editTextReportItem"
